WIP(v2.32): merge proptest_gen.rs into proptest_gen_mir, delete it
proptest_gen_mir was a 100% pure-delegation scaffold (let _ = mir;
calling proptest_gen::generate). Folded the entire proptest_gen.rs
implementation (~2.1K LoC: the per-handler arb_state / preservation /
invariant / guard / overflow / sequence sub-emitters + the record /
unit-sum composers + tests) into proptest_gen_mir.rs as the now-private
generate_impl, called by the public generate(mir, parsed, spec, out).

The harness emitters read ParsedSpec directly — they derive from
account / property / requires surface, not the effect-body Stmt IR, so
they correctly stay ParsedSpec-based; the public entry keeps Mir on its
signature for dispatch uniformity.

Repointed the cross-module helper users (rust_codegen_util, kani_mir →
proptest_gen_mir::default_value_for_field) and regen_drift's proptest
regen (→ proptest_gen_mir::generate, parse+lower locally). Removed
`mod proptest_gen;`.

proptest_snapshot byte-identical (6/6); 980 tests pass; fmt + clippy
clean. (-1 test = the trivial scaffold smoke dropped in the merge.)
